Reworking to use new sdc rabbit library
