+++ startup.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-python3 -m app.consumer
+python3 -m app.main
